pyproject.toml
[tool.poetry]
name = "optimus"
version = "0.1.0"
description = ""
authors = ["cschneuw <christelle.schneuwly.diaz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.2.3"
scipy = "^1.15.2"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
missingno = "^0.5.2"
miceforest = "^6.0.3"
xgboost = "^2.1.4"
pytorch-tabnet = "^4.1.0"
nilearn = "^0.11.1"
jupyter = "^1.1.1"
cmap = "^0.5.0"
matplotlib-venn = "^1.1.2"
MissForest = "^4.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
